% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{dtm_trim}
\alias{dtm_trim}
\title{Trim a fastDtm Object}
\usage{
dtm_trim(dtm, min_docfreq = 1, min_termfreq = 1)
}
\arguments{
\item{dtm}{A fastDtm object.}

\item{min_docfreq}{Minimum document frequency.}

\item{min_termfreq}{Minimum corpus term frequency.}
}
\value{
A trimmed fastDtm object.
}
\description{
Removes infrequent terms from a fastDtm object.
}
