% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\docType{class}
\name{fastDtm-class}
\alias{fastDtm-class}
\title{fastDtm Class}
\description{
An S4 class extending \code{\linkS4class{dgCMatrix}} for storing
document-term matrices together with vocabulary statistics and
document-level metadata.
}
\section{Slots}{

\describe{
\item{\code{vocabulary}}{Character vector of unique terms.}

\item{\code{term_frequency}}{Corpus-wide term frequencies.}

\item{\code{document_frequency}}{Number of documents containing each term.}

\item{\code{docvars}}{Data frame of document-level metadata.}
}}

