% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FED_Minutes.R
\docType{data}
\name{FED_Minutes}
\alias{FED_Minutes}
\title{Federal Reserve Meeting Minutes}
\format{
A data frame with 249 observations and 2 variables:
\describe{
\item{Date}{The publication date of the meeting minutes.}
\item{Text}{The full text of the published FOMC meeting minutes.}
}
}
\source{
Federal Reserve Board of Governors,
\url{https://www.federalreserve.gov/monetarypolicy/fomccalendars.htm}
}
\usage{
FED_Minutes
}
\description{
A dataset containing the published minutes of meetings of the
Federal Open Market Committee (FOMC) from March 1994 to November 2024.
}
\keyword{datasets}
