% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opttext_package.R
\docType{package}
\name{opttext-package}
\alias{opttext}
\alias{opttext-package}
\title{opttext: Optimised Text Preprocessing}
\description{
High-performance text preprocessing functions for NLP workflows,
implemented with C++ (via \pkg{Rcpp}) and parallel computing
(via \pkg{RcppParallel}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VikyRackova/opttext}
  \item Report bugs at \url{https://github.com/VikyRackova/opttext/issues}
}

}
\author{
\strong{Maintainer}: Viktoria Račková \email{viki.rackova@gmail.com}

}
\keyword{internal}
