% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-preprocessing.R
\name{remove_numbers}
\alias{remove_numbers}
\title{Remove Numbers}
\usage{
remove_numbers(texts)
}
\arguments{
\item{texts}{Character vector.}
}
\value{
Character vector.
}
\description{
Removes numeric characters from text.
}
\examples{
remove_numbers("Inflation increased by 5 percent")

}
