% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-preprocessing.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Remove Punctuation}
\usage{
remove_punctuation(texts)
}
\arguments{
\item{texts}{Character vector.}
}
\value{
Character vector.
}
\description{
Removes all Unicode punctuation characters from text.
}
\examples{
remove_punctuation("Inflation, growth, and employment.")

}
