% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The_Times.R
\docType{data}
\name{The_Times}
\alias{The_Times}
\title{The Times Newspaper Coverage of Government and Public Administration}
\format{
A data frame with 179,760 observations and 2 variables:
\describe{
\item{Date}{The publication date of the article.}
\item{Text}{The full text of the article.}
}
}
\source{
LexisNexis archive of \emph{The Times}
}
\usage{
The_Times
}
\description{
A dataset containing articles published in \emph{The Times}
that discuss government and public administration.
The dataset covers the period from January 2020 to December 2025.
}
\keyword{datasets}
