% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-preprocessing.R
\name{squish_whitespace}
\alias{squish_whitespace}
\title{Squish Whitespace}
\usage{
squish_whitespace(texts)
}
\arguments{
\item{texts}{Character vector.}
}
\value{
Character vector.
}
\description{
Removes leading and trailing whitespace and converts repeated
whitespace to a single space.
}
\examples{
squish_whitespace("inflation    growth   employment")

}
